+++ src/browser/tab/mod.rs
@@ -530,8 +530,8 @@ impl<'a> Tab {
     /// # use failure::Fallible;
     /// # fn main() -> Fallible<()> {
     /// #
-    /// use headless_chrome::{protocol::page::ScreenshotFormat, Browser, LaunchOptionsBuilder};
-    /// let browser = Browser::new(LaunchOptionsBuilder::default().build().unwrap())?;
+    /// use headless_chrome::{protocol::page::ScreenshotFormat, Browser, LaunchOptions};
+    /// let browser = Browser::new(LaunchOptions::default_builder().build().unwrap())?;
     /// let tab = browser.wait_for_initial_tab()?;
     /// let viewport = tab.navigate_to("https://en.wikipedia.org/wiki/WebKit")?
     ///     .wait_for_element("#mw-content-text > div > table.infobox.vevent")?