+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 * `LaunchOptions::default_builder()` for easier access to LaunchOptionsBuilder (which, because it's created via `derive(Builder)` is hard for editors to deal with
+* `Element::get_inner_text()`
 ### Removed
 ### Changed
 * Move env_logger to dev dependencies